python_code/api/requirements.txt
numpy==1.26.4
pandas==2.3.2
python-dotenv==1.0.1
openai==1.50.2
runpod==1.7.1
pinecone==5.3.1
httpx==0.27.0